oem-ami: HACK: add zeroconf route for openstack
Issue #184

Problem: Unable to log into instance with pubkey auth, as ssh key
retrieval can't hit the metadata service.

Environment: coreos 54.0.0 openstack image in openstack grizzly, quantum
networking with openvswitch, provider network (i.e. bridged to real vlan
on network).

Upon further examination: our setup might be totally broken, but
instances can't reach the metadata service when routing to the default
gateway.  Our instances MUST have a zeroconf route to the link-local
network.

This is NOT meant to be merged. This is really the ami oem, which I
understand is just converted to qcow2 for openstack post-build, and may
break in AWS for all I know.