Hack to enable zeroconf route.
...so ec2 metadata service is reachable via link-local in Openstack with
provider networks.

See coreos/coreos-overlay/issues/184 for more details.

This is what I use in our Openstack Havana setup on provider networks,
where 169.254.169.254 is reachable only via link-local, not the gateway.
FWIW, I just crack the image after download and shim this in
/usr/share/oem/cloud-config.yml, this commit is just for better
visibility.

This is not intended to be a general purpose solution, with the
hardcoded interface name, etc etc.  Also have no idea if this is an
appropriate place in the system lifecycle to initialize this.

Maybe a more general solution is this: after net init, arping for
169.254.169.254 on the "primary" interface (generally whatever interface
has the default route).  Add the zeroconf route iff the arping succeeds.